+++ azure-storage-common/azure/storage/common/_deserialization.py
@@ -22,6 +22,7 @@ from .models import (
     GeoReplication,
     ServiceStats,
     DeleteRetentionPolicy,
+    StaticWebsite,
 )
 
 
@@ -253,6 +254,11 @@ def _convert_xml_to_service_properties(response):
              <Enabled>true|false</Enabled>
              <Days>number-of-days</Days>
         </DeleteRetentionPolicy>
+        <StaticWebsite>
+            <Enabled>true|false</Enabled>
+            <IndexDocument></IndexDocument>
+            <ErrorDocument404Path></ErrorDocument404Path>
+        </StaticWebsite>
     </StorageServiceProperties>
     '''
     if response is None or response.body is None:
@@ -321,6 +327,20 @@ def _convert_xml_to_service_properties(response):
 
         if policy_enabled:
             service_properties.delete_retention_policy.days = int(delete_retention_policy_element.find('Days').text)
+
+    # StaticWebsite
+    static_website_element = service_properties_element.find('StaticWebsite')
+    if static_website_element is not None:
+        service_properties.static_website = StaticWebsite()
+        service_properties.static_website.enabled = _bool(static_website_element.find('Enabled').text)
+
+        index_document_element = static_website_element.find('IndexDocument')
+        if index_document_element is not None:
+            service_properties.static_website.index_document = index_document_element.text
+
+        error_document_element = static_website_element.find('ErrorDocument404Path')
+        if error_document_element is not None:
+            service_properties.static_website.error_document_404_path = error_document_element.text
 
     return service_properties