+++ azure-storage-queue/azure/storage/queue/_constants.py
@@ -8,4 +8,4 @@ __author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
 __version__ = '1.2.0rc1'
 
 # x-ms-version for storage service.
-X_MS_VERSION = '2017-11-09'
+X_MS_VERSION = '2018-03-28'