+++ azure-storage-blob/ChangeLog.md
@@ -2,6 +2,10 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
+## Version XX.XX.XX:
+
+- Added support for static website service properties.
+
 ## Version 1.2.0rc1:
 
 - Support for 2017-11-09 REST version. Please see our REST API documentation and blog for information about the related added features.